perf(memory): shrink Value/AST, reclaim boot interpreter, honor SOLI_WORKERS
Reduce per-worker server memory, which scales with worker count (each worker
thread holds its own parsed app + builtins; Rc values are !Send).

Core type sizes (locked by compile-time size_of guards):
- Value 64 -> 24 bytes: NativeFunction is a newtype over Rc<..> (Deref, so
  ~464 call sites unchanged) and Method is boxed.
- Expr 144 -> 80, Stmt 360 -> 200: Span fields usize -> u32 (readers use
  span.*_usize() accessors), fat StmtKind decls + Lambda return type boxed.
  SLAST bundle format unaffected (MessagePack encodes ints by value).

Serve:
- Boot interpreter (seeds shared registries only) is reclaimed after boot
  instead of parked for the process lifetime; uses new_for_serve() (no test
  builtins). Clears view helpers / i18n cache / model classes then breaks the
  global-env<->closure cycle before drop.
- SOLI_JOB_VIEW_HELPERS=0 skips locale tables in job interpreters;
  DEFAULT_JOB_WORKERS 2 -> 1.
- fix: `soli serve` now honors SOLI_WORKERS (the CLI ignored it, reading only
  --workers/CPU cores); --workers still overrides.

+++ src/lint/mod.rs
@@ -67,7 +67,7 @@ impl Linter {
 
         let suppressions = suppress::collect_suppressions(&self.source);
         self.diagnostics
-            .retain(|d| !suppressions.suppresses(d.span.line as u32, d.rule));
+            .retain(|d| !suppressions.suppresses(d.span.line, d.rule));
 
         self.diagnostics
             .sort_by_key(|d| (d.span.line, d.span.column));
